remove and uninstall old tooltip
