+++ packages/client/package.json
@@ -41,7 +41,6 @@
     "react-table": "^7.5.0",
     "react-toastify": "^6.0.9",
     "react-window": "^1.8.7",
-    "reactjs-popup": "^2.0.4",
     "redux": "^4.0.5",
     "redux-devtools-extension": "^2.13.8",
     "redux-thunk": "^2.3.0",
